fix(lib|php|package|vcs): remove doom's `defadvice!` macro
The Radian macro, which I've renamed to `def-advice!`, will replace `defadvice!`
when possible. My primary reason is that `defadvice!` stupidly shuffles around
the order of arguments in such a way that there is no possible way to remember
when comparing with `advice-add`. The Radian macro doesn't do such things.

Note that `def-advice!` does not support adding to multiple places via a list.
For that, use a regular defun and `advice-add`.