env|eat|git: ensure shell/term commands use with-editor
Sometimes, I am unable to make a Git commit with Magit because it is waiting for
some process to "close the editor". I'm not sure what causes this, but the last
time it happened, I had a running `eat` terminal process. Killing that process
did not resolve the issue, however.